requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Notion API Client
notion-client==2.2.1

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# HTML/XML Parsing
beautifulsoup4==4.12.2
lxml==5.1.0
html2text==2020.1.16
readability-lxml==0.8.1

# Database Integration
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Security & Encryption
cryptography==41.0.7
flask-limiter==3.5.0  # Rate limiting
flask-talisman==1.1.0  # Security headers
flask-jwt-extended==4.6.0  # JWT authentication

# Validation
marshmallow==3.20.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
faker==22.0.0

# Utilities
python-dateutil==2.8.2

# Report Generation
openpyxl==3.1.2

# API Documentation
flask-swagger-ui==4.11.1
apispec==6.3.1
PyYAML==6.0.1

# Monitoring & Metrics
prometheus-client==0.19.0
psutil==5.9.6

# Background Task Processing
redis==5.0.1
rq==1.15.1
rq-dashboard==0.6.1
rq-scheduler==0.13.1
fakeredis==2.20.1  # For testing without Redis server
